refactor(habitat): classify check protection by kind
