+++ tools/habitat-harness/src/service/model/check/policy/protected-zone-path-actions.policy.ts
@@ -1,6 +1,9 @@
 import path from "node:path";
 import { Value } from "typebox/value";
-import { type StagedMutationPath, StagedMutationPathSchema } from "./schema.js";
+import {
+  type StagedMutationPath,
+  StagedMutationPathSchema,
+} from "../dto/protected-zone.schema.js";
 
 export function stagedPathsFromNameStatus(output: string): StagedMutationPath[] {
   const tokens = output.split("\0").filter(Boolean);
