--- tools/habitat-harness/src/service/model/check/protection-policy/index.ts
@@ -1,20 +1,0 @@
-export {
-  hostApplyGateDecision,
-  hostAuthoringBoundaryState,
-  hostGeneratedSurfaceDeclarations,
-  hostProjectSupportDecision,
-  hostSurfaceDecisionForGeneratedZone,
-  hostSurfaceDecisionForPath,
-  hostSurfaceDecisionForScanRoot,
-  matchesHostMatcher,
-  renderHostRecoveryInstruction,
-} from "./decisions.js";
-export { defaultHostPolicyDocument } from "./declarations.js";
-export * from "./schema.js";
-export {
-  defaultHostPolicyState,
-  missingHostPolicyState,
-  parseHostPolicyDocument,
-  readHostPolicyState,
-  unavailableHostPolicyState,
-} from "./state.js";
