+++ tools/habitat-harness/test/lib/host-policy.test.ts
@@ -9,7 +9,7 @@ import {
   missingHostPolicyState,
   readHostPolicyState,
   unavailableHostPolicyState,
-} from "@internal/habitat-harness/service/model/check/protection-policy";
+} from "@internal/habitat-harness/service/model/check/index";
 import { describe, expect, test } from "vitest";
 
 describe("host policy boundary", () => {
